reorganize records folder to use css next to component organization strategy
