+++ apps/genomicsdb/app/record/layout.tsx
@@ -1,10 +1,8 @@
-import styles from "@/components/records/styles/record.module.css";
-
 export default async function RecordPageLayout({ children }: { children: React.ReactNode }) {
     return (
-        <div className={`${styles.pageContainer} ${styles.pageContainerCapped}`}>
-            <div className={styles.content}>
-                <div className={styles.contentSection}>{children}</div>
+        <div className="page-container page-container-capped">
+            <div className="content">
+                <div className="content-section">{children}</div>
             </div>
         </div>
     );
