+++ apps/genomicsdb/components/records/MostSevereConsecuenceCard/most-severe-consequence-card.module.css
