+++ apps/genomicsdb/components/records/RecordPage/record-page.module.css
@@ -18,8 +18,6 @@
     margin-left: 52px;
 }
 
-
-
 @media (max-width: 768px) {
     /* Main content always full width */
     .record-content {
