+++ apps/genomicsdb/components/records/RecordSidebar/record-sidebar.module.css
@@ -278,8 +278,8 @@
     cursor: pointer;
     align-items: center;
     justify-content: center;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, .3);
-    transition: background .15s;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+    transition: background 0.15s;
 }
 
 .sidebar-fab:hover {
@@ -291,12 +291,11 @@
     display: none;
     position: fixed;
     inset: 0;
-    background: rgba(0, 0, 0, .4);
+    background: rgba(0, 0, 0, 0.4);
     z-index: 99;
 }
 
 @media (max-width: 768px) {
-
     /* Show FAB and overlay */
     .sidebar-fab {
         display: flex;
@@ -309,7 +308,7 @@
     /* Sidebar: hidden off-screen by default */
     .sidebar {
         transform: translateX(-100%);
-        transition: transform .25s cubic-bezier(.4, 0, .2, 1);
+        transition: transform 0.25s cubic-bezier(0.4, 0, 0.2, 1);
         width: 100% !important;
         /* override collapsed/expanded desktop widths */
         top: 0;
