+++ apps/genomicsdb/styles/layout.css
@@ -36,6 +36,15 @@
     padding: 1.5rem;
 }
 
+.content {
+    overflow: auto;
+    width: 100%; /* EGA -temp w/out sidebar */
+}
+
+.content-section {
+    margin-top: 0;
+}
+
 .content-header {
     display: flex;
     justify-content: space-between;
@@ -54,4 +63,15 @@
 .content-subtitle {
     color: var(--text-secondary);
     margin-top: 0.5rem;
+}
+
+.page-container {
+    display: flex;
+
+    min-height: calc(100vh - var(--header-height));
+    margin-top: 0;
+}
+
+.page-container-capped {
+    height: calc(100vh - var(--header-height));
 }
